services/user/requirements.txt
fastapi==0.111.1
mysql-connector-python==9.0.0
boto3==1.34.150